+++ sdk/metric/periodic_reader.go
@@ -31,6 +31,7 @@ type periodicReaderConfig struct {
 	timeout                  time.Duration
 	producers                []Producer
 	cardinalityLimitSelector CardinalityLimitSelector
+	metricFilter             metricFilter
 }
 
 // newPeriodicReaderConfig returns a periodicReaderConfig configured with
@@ -41,8 +42,15 @@ func newPeriodicReaderConfig(options []PeriodicReaderOption) periodicReaderConfi
 		timeout:                  envDuration(envTimeout, defaultTimeout),
 		cardinalityLimitSelector: defaultCardinalityLimitSelector,
 	}
-	for _, o := range options {
-		c = o.applyPeriodic(c)
+	for _, opt := range options {
+		if o, ok := opt.(metricFilter); ok {
+			c.metricFilter = o
+			continue
+		}
+		if _, ok := opt.(experimentalOption); ok {
+			continue
+		}
+		c = opt.applyPeriodic(c)
 	}
 	return c
 }
@@ -120,6 +128,7 @@ func NewPeriodicReader(exporter Exporter, options ...PeriodicReaderOption) *Peri
 		flushCh:                  make(chan chan error),
 		cancel:                   cancel,
 		done:                     make(chan struct{}),
+		metricFilter:             conf.metricFilter,
 		cardinalityLimitSelector: conf.cardinalityLimitSelector,
 		rmPool: sync.Pool{
 			New: func() any {
@@ -181,6 +190,8 @@ type PeriodicReader struct {
 	cardinalityLimitSelector CardinalityLimitSelector
 
 	inst *observ.Instrumentation
+
+	metricFilter metricFilter
 }
 
 // Compile time check the periodicReader implements Reader and is comparable.
@@ -235,6 +246,11 @@ func (r *PeriodicReader) aggregation(
 // cardinalityLimit returns the cardinality limit for kind.
 func (r *PeriodicReader) cardinalityLimit(kind InstrumentKind) (int, bool) {
 	return r.cardinalityLimitSelector(kind)
+}
+
+// metricFilter returns the MetricFilter configured for this reader.
+func (r *PeriodicReader) getMetricFilter() metricFilter {
+	return r.metricFilter
 }
 
 // collectAndExport gather all metric data related to the periodicReader r from